third_party/blink/web_tests/platform/mac/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/template.window-expected.txt

This is a testharness.js-based test.
[FAIL] Serializing a ShadowRoot with a null registry with a scoped registry host document
  assert_equals: expected "<template shadowrootmode=\\"closed\\" shadowrootserializable=\\"\\" shadowrootcustomelementregistry=\\"\\"></template>" but got "<template shadowrootmode=\\"closed\\" shadowrootserializable=\\"\\"></template>"
Harness: the test ran to completion.
